requirements.txt
tensorflow-cpu==2.12.0
pyqt5
pillow
opencv-python
matplotlib